+++ spring-expression/src/test/java/org/springframework/expression/spel/AbstractExpressionTests.java
@@ -177,7 +177,7 @@ public abstract class AbstractExpressionTests {
 			assertThat(ex.getMessageCode()).isEqualTo(expectedMessage);
 			if (!ObjectUtils.isEmpty(otherProperties)) {
 				// first one is expected position of the error within the string
-				int pos = ((Integer) otherProperties[0]).intValue();
+				int pos = (Integer) otherProperties[0];
 				assertThat(ex.getPosition()).as("position").isEqualTo(pos);
 				if (otherProperties.length > 1) {
 					// Check inserts match
@@ -207,7 +207,7 @@ public abstract class AbstractExpressionTests {
 			assertThat(ex.getMessageCode()).isEqualTo(expectedMessage);
 			if (otherProperties != null && otherProperties.length != 0) {
 				// first one is expected position of the error within the string
-				int pos = ((Integer) otherProperties[0]).intValue();
+				int pos = (Integer) otherProperties[0];
 				assertThat(pos).as("reported position").isEqualTo(pos);
 				if (otherProperties.length > 1) {
 					// Check inserts match
